update for model file breakout
